wip: cart context state + list cart items
