+++ frontend/src/carts/cartContext.tsx
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import { CartResponse } from "./api";
+
+export const CartItemsContext = createContext<CartResponse | undefined>(
+  undefined
+);
